# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TeamTime API",
    "version": "1.0.0",
    "description": "API REST para gestión de tiempo, proyectos y recursos humanos con sistema de roles jerárquico",
    "contact": {
      "name": "TeamTime Development Team",
      "email": "desarrollo@teamtime.com",
      "url": "https://github.com/your-org/teamtime-backend"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.teamtime.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Descripción del error"
          },
          "error": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "object",
            "properties": {
              "field": {
                "type": "string",
                "example": "email"
              },
              "code": {
                "type": "string",
                "example": "INVALID_FORMAT"
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "user-123"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@ejemplo.com"
          },
          "firstName": {
            "type": "string",
            "example": "Juan"
          },
          "lastName": {
            "type": "string",
            "example": "Pérez"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMINISTRADOR",
              "COORDINADOR",
              "COLABORADOR"
            ],
            "example": "COLABORADOR"
          },
          "areaId": {
            "type": "string",
            "format": "uuid",
            "example": "area-456"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-07-04T10:30:00.000Z"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Token de autenticación inválido o faltante",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Token de autenticación requerido",
              "error": "UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Permisos insuficientes para realizar la operación",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "No tienes permisos para realizar esta acción",
              "error": "FORBIDDEN"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Recurso no encontrado",
              "error": "NOT_FOUND"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Error de validación en los datos enviados",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "success": false,
              "message": "Error de validación",
              "error": "VALIDATION_ERROR",
              "details": {
                "field": "email",
                "code": "INVALID_FORMAT"
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/areas/{id}/users": {
      "get": {
        "summary": "Obtener usuarios de un área",
        "description": "Retorna la lista de usuarios pertenecientes a un área específica.",
        "tags": [
          "Áreas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID del área"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por usuarios activos/inactivos"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuarios del área obtenidos exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuarios del área obtenidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 25
                            },
                            "pages": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "summary": "Crear nuevo usuario",
        "description": "Permite a administradores y coordinadores crear nuevos usuarios en el sistema.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "firstName",
                  "lastName",
                  "role",
                  "areaId"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "nuevo.usuario@ejemplo.com"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "ContraseñaSegura123"
                  },
                  "firstName": {
                    "type": "string",
                    "example": "Carlos"
                  },
                  "lastName": {
                    "type": "string",
                    "example": "Rodríguez"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "ADMINISTRADOR",
                      "COORDINADOR",
                      "COLABORADOR"
                    ],
                    "example": "COLABORADOR"
                  },
                  "areaId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "area-123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuario creado exitosamente"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "get": {
        "summary": "Obtener lista de usuarios",
        "description": "Retorna una lista paginada de usuarios con filtros opcionales. Los coordinadores solo ven usuarios de su área.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Número de página"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Cantidad de elementos por página"
          },
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "ADMINISTRADOR",
                "COORDINADOR",
                "COLABORADOR"
              ]
            },
            "description": "Filtrar por rol"
          },
          {
            "in": "query",
            "name": "areaId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filtrar por área"
          },
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por estado activo/inactivo"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Buscar por nombre o email"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Usuarios obtenidos exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/User"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "total": {
                              "type": "integer",
                              "example": 25
                            },
                            "pages": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/users/stats": {
      "get": {
        "summary": "Obtener estadísticas de usuarios",
        "description": "Retorna estadísticas generales sobre los usuarios del sistema. Solo disponible para administradores.",
        "tags": [
          "Usuarios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Estadísticas obtenidas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Estadísticas obtenidas exitosamente"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalUsers": {
                          "type": "integer",
                          "example": 150
                        },
                        "activeUsers": {
                          "type": "integer",
                          "example": 142
                        },
                        "inactiveUsers": {
                          "type": "integer",
                          "example": 8
                        },
                        "usersByRole": {
                          "type": "object",
                          "properties": {
                            "ADMINISTRADOR": {
                              "type": "integer",
                              "example": 3
                            },
                            "COORDINADOR": {
                              "type": "integer",
                              "example": 15
                            },
                            "COLABORADOR": {
                              "type": "integer",
                              "example": 132
                            }
                          }
                        },
                        "usersByArea": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "areaName": {
                                "type": "string",
                                "example": "Desarrollo"
                              },
                              "userCount": {
                                "type": "integer",
                                "example": 25
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    }
  }
}